* release: (4879 commits)
Update: Grunt
Update: Grunt string-replace
Update: Version name update
Update: Language file update
added change log
PROD-5978 - update code to run migration once on the update
PROD-6172 - pull the api changes from the repo
PROD-6172 - pull the API changes from repo 2.4.50
PROD-5342 : Updated db version and migration function
[PROD-2838] grunt
PROD-6275 - clear api cache first before to flush object cache
PROD-6275 - change setup value to true to bool check
PROD-6275 - phpcbf fixes
PROD-6275 - updated callback name
PROD-6275 - updated doc block
PROD-6275 - fix the issue with the bb_activity_like_reaction background jobs flooding
Update: Grunt
Update: Grunt string-replace
Update: Version name update
grunt
...

# Conflicts:
#	src/bp-forums/admin/metaboxes.php
#	src/bp-forums/admin/tools.php
#	src/bp-forums/forums/functions.php

+++ src/bp-core/bp-core-tools-default-data.php
@@ -243,11 +243,13 @@ function bp_dd_delete_dummy_forum() {
 	/**
 	 * Delete Forums
 	 */
-	$forums = bp_get_option( 'bp_dd_imported_forum_ids' );
-	if ( ! empty( $forums ) ) {
-		foreach ( (array) $forums as $forum_id ) {
-			wp_delete_post( $forum_id );
-			bbp_delete_forum( $forum_id );
+	if ( bp_is_active( 'forums' ) ) {
+		$forums = bp_get_option( 'bp_dd_imported_forum_ids' );
+		if ( ! empty( $forums ) ) {
+			foreach ( (array) $forums as $forum_id ) {
+				wp_delete_post( $forum_id );
+				bbp_delete_forum( $forum_id );
+			}
 		}
 	}
 }
@@ -687,6 +689,34 @@ function bp_dd_import_users() {
 				'user_pass'       => wp_generate_password( 8, false ),
 			)
 		);
+
+		if ( is_wp_error( $user_id ) ) {
+
+			// If multisite and user already exist, then add user to the current blog.
+			if (
+				is_multisite() &&
+				'existing_user_login' === $user_id->get_error_code()
+			) {
+				$userdata = get_user_by( 'login', $user['login'] );
+
+				if ( ! empty( $userdata->ID ) ) {
+					$result = add_existing_user_to_blog(
+						array(
+							'user_id' => $userdata->ID,
+							'role'    => get_option( 'default_role' ),
+						)
+					);
+
+					// If user added to current blog then set this ID to $users.
+					if ( ! is_wp_error( $result ) ) {
+						$users[] = $userdata->ID;
+					}
+				}
+			}
+
+			// If found an errors then continue loop.
+			continue;
+		}
 
 		if ( bp_is_active( 'xprofile' ) ) {
 			xprofile_set_field_data( 1, $user_id, $user['display_name'] );
